requirements.txt
langchain==0.1.11
langchain-openai==0.1.3
langchain-community==0.0.38
streamlit==1.28.1
pdfplumber==0.10.3
PyPDF2==3.0.1
tavily-python==0.7.19
pandas==2.1.2
pydantic==2.5.0
python-dotenv==1.0.0
